# Speed up inference